requirements.txt
scikit-learn
fastapi
uvicorn
pydantic
sqlalchemy
scikit-surprise
pandas
pyodbc